Reorganize file structure and add margin above artist image
