+++ src/styles/global.scss
@@ -1,5 +1,5 @@
 @import './fonts.sass';
-@import '../node_modules/bulma/bulma.sass';
+@import '../../node_modules/bulma/bulma.sass';
 
 * {
   z-index: 1;
